Use CFGInliner plugin to extend functionality for inlining an object and appropriate method
Summary:
Builds a plugin for the Inliner that will inline a method body in place of an object value, and thus will redirect field iput and iget into moves, and does not remove the instruction which is controlling the placement of the inlined blocks.

This uses the data collected by the ClassInitCounter to control which field accesses are modified.

The inliner does not modify or remove any new_instance instructions or other origins of the object value, leaving this to the transform caller or dead code analysis depending on caller's choice.

Reviewed By: justinjhendrick

Differential Revision: D16420225

fbshipit-source-id: 498c452a64935e25a636b92efb2cc1633406076a