Automatically clean up dangling POSITIONS.
Summary:
`insert` can remove `MethodItemEntry`-s when the newly introduced sequence of
instructions ends in a `return` or a `throw`.  This includes `POSITION`s, which
could be acting as parents to other positions.

This diff has `insert` keep track of the positions it is about to delete, so it
can scan for their appearance in parent slots.

Reviewed By: justinjhendrick

Differential Revision: D18464885

fbshipit-source-id: cf7a804a972fb019ca8fc79d33235b905ca0eeb1