Do not desuper invoke-super on interface default methods
Summary:
As indicated in the 037 spec, `invoke-super` is also used to invoke default interface methods. Just in case if we are processing 037 code, we want to avoid rewriting invoke on default methods.

"In Dex files version 037 or later, if the method_id refers to an interface method, invoke-super is used to invoke the most specific, non-overridden version of that method defined on that interface. The same method restrictions hold as for invoke-virtual. In Dex files prior to version 037, having an interface method_id is illegal and undefined."

Reviewed By: int3

Differential Revision: D18586435

fbshipit-source-id: 6ccd84da87e445980023bf0a7e399bc3ad0e3315