Abstract over operation in test helper
Summary:
In the next diff I am going to employ a new kind of CFG mutation in
`RemoveUninstantiablesPass`.  In order to test it, this diff abstracts over the
operation that `EXPECT_CHANGE` applies before comparing the IR.

Most of this diff is noise from code formatting, so I have split it into its
own diff.

Reviewed By: NTillmann

Differential Revision: D18492244

fbshipit-source-id: 25a9732aafc4ba73abcade6b61c217020e8e39d3