Tell GDB where to find source files
Summary:
Add a flag to the python runner to specify a root for source files.  If it
exists, it is used by the debugger to augment its source search path.  This is
for compatibility with build systems that use relative source locations in
debug information for the binaries they produce.

Differential Revision: D18658015

fbshipit-source-id: 8839d899d49022551253bfbd8c109cf021d9e8bb