Refine virtual callsites based on Type Inference
Summary:
As part of the method reference binding reorganization, I want to relocated the virtual call site rebinding logic currently in ReBindVRefs into ResolveRefsPass. The reason is that they both try to bind virtual call site references to the most accurate callee possible. We want to do that before inlining and other optimizations before we rebind reference back to more general ones towards the end in ReBindRefs. Doing so simplifies our pass configs and keeping reference binding logic in two passes.

The change currently moves the logic in ReBindVRefs into ResolveRefs. So for each method we iterate the instructions twice. First we rebind method and field references using the existing logic. In the second iteration we do exactly what ReBindVRefs does today. We repeat ResolveRefs again right before the 2nd instance of ReBindVRefs we have in the config.

The clean up will come in the next diff to remove ReBindVRefs.

Reviewed By: int3, agampe

Differential Revision: D18605956

fbshipit-source-id: 534ae84e8dc358c5abce99e7a351fdeb97d68725