Print warnings on mergeable with no ctor
Summary: We didn't have check on whether an incoming mergeable class has at least one ctor or not. If such a mergeable exists we still merge them and adding the missing ctor back to it. The diff adds a warning message when we found such a mergeable, but the merging logic is unchanged.

Reviewed By: ssj933

Differential Revision: D18734388

fbshipit-source-id: e67dfa98e1323195f38750641e99a02dea623b08