LoopInfo framework and TestLoopPass
Summary:
This diff adds a basic loop detection framework based off of LLVM's LoopInfo
implementation. The focus of this diff is to provide a mechanism for creating
Loop objects, but does not attempt to create APIs that leverage these Loop
objects (will come in a future diff).

We first construct a weak topological ordering of the control
flow graph in order to get strongly connected components. We then iterate
through the SCC's to check whether or not they are valid loops.

Our loop definition matches LLVM's loop definition - a loop is a SCC with a
designated dominating header block.

TestLoopPass is implemented as a test pass to run on our apps. It currently
tells us how many loops exist.

Reviewed By: emma0303

Differential Revision: D18896799

fbshipit-source-id: 4f07a55f31c63a0af439192767ba9b5c9ebe474e