Make reverse postorder algorithm more generic and non-quadratic
Summary:
Resurrecting this diff that I did ages ago just because...

The original motivation for this diff was to run the dominators algorithm (which in turn uses postorder sorting) on the reachability graph. Didn't turn out to be very useful in practice, so I didn't push ahead with landing it then, but this is still nice as a code cleanup.

Additionally, the old implementation of the postorder sort was quadratic in the number of block successors. Didn't really notice a perf issue in practice but it's nice to fix.

The only functional difference is that the new implementation doesn't include unreachable blocks in its output. Unfortunately, last I checked, RemoveBuilders relies on them being included for correct behavior, so the old implementation is still kept around with a `_deprecated` suffix.

Reviewed By: ssj933

Differential Revision: D13564148

fbshipit-source-id: 50a1a9c25a811a3156c7c0cfc101fc012aee9020