Resolve virtual method reference including non-exact match
Summary:
`RebindVRefs` resolves virtual call site method reference based on locally inferred DexType and method override graph. However, it does this in a limited fashion which only looks at overriding method defined on the inferred DexType. This is sometimes not enough.

In addition, up to this point we don't touch external method references. We can resolve a large number of external method references to its concrete implementation to minimize method reference count. However, we want to do this with extra caution. We don't want to touch Android support library and Android X that deal with cross version compatibilities. We are likely to emit method references that don't exist in older versions of Android. On top of that, we still have a list of external method prefixes that we should never resolve method reference to. The variation of those methods on older Android versions will cause method not found error.

Reviewed By: int3

Differential Revision: D18660529

fbshipit-source-id: f8fac2c93a8eb965e003cd96554ddbe42d20f8d7