API to clear pending changes
Summary:
Calling this function clears all the changes that have accumulated in the
mutation so far.

This also fixes a potential memory leak in the API where if a change's anchor
gets removed, it is dropped, but previously its instructions were not being
freed.  It is not in general safe to do this in the destructor, because a
`ChangeSet` may be holding on to instructions whose ownership has been
transferred to an `IRCode` when it is being destroyed.

Reviewed By: rybalkinsd

Differential Revision: D19033091

fbshipit-source-id: 4554bc3bbcb6a283081afa34a39561b5c389c120