LoopInfo API
Summary:
Adds new API to LoopInfo.

This diff adds forward and reverse iterators to Loop and LoopInfo. For Loop you
can now iterate over the blocks in the loop as well as the subloops contained
by the loop.

This diff also adds the function `LoopInfo::get_loop_for(cfg::Block*)`, which
returns the innermost loop that contains the block.

Reviewed By: yuxuanchen1997

Differential Revision: D19185530

fbshipit-source-id: 58bc4923591ca8b878d243ecda6ef50e567cc834