Fix in ConstantUses, second attempt
Summary:
This fixes a bug in an silly wrong condition:
```
   (opcode == OPCODE_APUT || OPCODE_APUT_WIDE)
```

The seemingly obvious way of fixing this would be to turn that line into...
```
   (opcode == OPCODE_APUT || opcode == OPCODE_APUT_WIDE)
```

However, that causes test failures. Here's what's happening:
- The buggy `OPCODE_APUT_WIDE` is equivalent to `true`, and thus we would *always* compute type inference information when any const value is used anywhere.
- Historically, it was only needed (i.e. the computed `m_type_inference` would only get accessed) in the cases of `APUT`/`APUT_WIDE`
- Fixing the condition by adding `opcode ==` would still cause that code to work.
- But some other, later added code, would break, as `m_type_inference` is also now used in the context of `IF_EQ` and `IF_NE` opcodes.

So the true fix is to compute type information in all of those cases.

I added a comment explanation how this is linked in thee code.

A nice side effect of this change should be a general speed-up, as we'll only compute type inference information now when it's actually needed.

Reviewed By: amnn

Differential Revision: D19254056

fbshipit-source-id: 33c9f9483b05e819e00cde52d3e22b61f7ea8524