Avoid unnecessary code traversals
Summary:
A cpu profile of TypeErasurePass revealed that we were spending a lot of cycles in `update_call_refs_simple`. After some research, I found that most of the invoke targets we updated were still the same as before. We were traversing the entire scope to make no changes. The `MethodDedup` service was adding identity entries into the `old_to_new_callee` map.

In this diff:
1) Don't add identity entries to the map
2) Skip the traversal if the incoming map is empty
3) Clean up the TRACE level usage a bit.

Reviewed By: thezhangwei

Differential Revision: D19278492

fbshipit-source-id: 2c6324db7cb67ec254dcbe5014ecd85ca2f8d60b