Removing support for non-name-based locators, removing dex file ordering dependency from class-loader
Summary:
Using magic locators, we can remove the existing ordering dependency on dex files and the complex store ordering logic from the class loader.
However, for that to work, we need to ensure that magic locators are always present, which comes with a few Redex changes (replicated across stable/unstable/staging).

Reviewed By: tophyr

Differential Revision: D18826164

fbshipit-source-id: 28da59519e23b928466c0551f6fd954f2592e883