Memoize comparator in profiled method order sort
Summary: I profiled `redex_backend` and found that over 50% of its time was spent in `stable_sort` for `sort_dexmethod_emitlist_profiled_order`. Most of that time was spent building full deobfuscated names to lookup in a map. That's a very expensive operation to do `O(n * logn)` times. Caching the results of the weight computation gives us a huge speedup!

Reviewed By: agampe, Feng23

Differential Revision: D19324583

fbshipit-source-id: f9f2d129fc8d6fbdd61f32f4d806672e630852d4