Rebind method reference to externals
Summary:
We used to rebind a number of method refs to external ones on JDK classes in a hard coded way. We can generalize the logic by traversing up the virtual scope, and the logic already exists in the pass. In addition, we can expand to rebind to more external methods to reduce method refs metadata.

- Generalized external method rebind logic to reuse the same visible top method lookup
- Rebind more external methods
- Do not rebind to methods listed in exclusion list; We know these classes have methods don't exist on older API versions.
- Do not rebind to external methods for call sites in support libraries (including Android X). However, we do rebind to JDK classes, since we know it's safe to do so.
- Rebind to external can be disabled by a flag which is off by default for OSS users.

Reviewed By: NTillmann

Differential Revision: D19266057

fbshipit-source-id: d4e89a458db1d5340cc3c68a8b9ff93d2807342b