Disable RebindVRefs
Summary:
I have relocated the logic in RebindVRefs to the revised ResolveRefsPass, and extended it to cover more cases. But I left RebindVRefs where it was before things become more stable.

We added more control on method ref rebinding to externals in ResolveRefs and disabled rebinding to externals for ICS builds. The reason is that FB4A ICS builds' min_sdk is 16 and the current target sdk is 28. There's quite a big gap between the two. It's hard to prevent an incorrect rebind from happening when running on older devices.

RebindVRefs doesn't handle rebind to externals. We have only enabled it for FB4A. It happens to be working so far but it's risky to keep doing so. I suggest we disabled it and let ResolveRefs and ReBindRefs handle all the method ref rebinding. That means we have a small regression for FB4A ICS builds (+51K). D19380499 can recoup a part of that (-22KB). So overall it's not a big deal, I think.

Reviewed By: int3

Differential Revision: D19402797

fbshipit-source-id: ad272128f7dae46babda272d1c800c9a05ea8380