Fix MallocDebug thread leak
Summary: Because the MallocDebug design is thread-local, assumedly to improve allocation performance slightly or avoid a complex design, remaining `Block`s must be deallocated when a MallocDebug instance is destroyed (i.e., when a thread exits). The Redex WorkQueue implementation always spawns new threads, so this actually happens a lot.

Reviewed By: amnn

Differential Revision: D19566045

fbshipit-source-id: 658b70a47a52dd93c2ff043fa8ef8fd47563b60a