Adding another interface to ControlFlowGraph::split_blocks()
Summary:
The new A/B experimentation infra needs a way to split a block's prologue from the rest of the block. We already have a method to split blocks, but it only accepts a `cfg::InstructionIterator`, which can only be an opcode (whereas we need to split specifically on a `DBG_SET_PROLOGUE_END` mie, which is not an opcode).

In this diff we create a new interface to `split_blocks()` which supports an IRList iterator as input. It uses the current implementation and is as safe as the existing one.

Reviewed By: justinjhendrick

Differential Revision: D19194526

fbshipit-source-id: 152574ff7d0a472383e8d39b5bbd7a7d5b4b09cf