Correct behaviour of splitting before
Summary:
Splitting the cfg before an instruction should actually always spawn a new block before the given instruction, where then the
given instruction is the first instruction. With a new block, this preserves the control flow into the instruction when it was
the first instruction previously and provides the right position for the inlined cfg blocks to fit between the instruction and
the previous instructions.

Reviewed By: justinjhendrick

Differential Revision: D19496742

fbshipit-source-id: f98738e7b42a51ad24ac9b3fdf1c2b7a02d067f9