Relocate FrameworkApi to libredex
Summary:
In ResolveRefs and RebindRefs we need to parse the min_sdk API file to check whether an external method binding exists in the SDK or not.
We repeat this process every time when we run these passes. This setup increased Redex running time by about 20 seconds.

This change moved the parsing of the API file into ConfigFile, and we do it lazily only once. The change requires moving a part of the logic from ApiLevelsUtils into libredex.

Reviewed By: emma0303

Differential Revision: D19633398

fbshipit-source-id: 074c65191a3dc72a10e82c04a1872354ae03a856