Reduce (common) size of `ReferencedState`
Summary:
In general reduce the size by using bitfields for the slow of flags.

As it is more common to not record keep reasons, make the mutex and map, which are quite expensive, indirections.

This reduces the size of `ReferencedState` to 32B, which has a knock-on effect because we have so many `DexMethod`s that directly include this: it reduces the size of `DexMethod` to 176B (from 136B/280B).

Reviewed By: int3, justinjhendrick

Differential Revision: D19654081

fbshipit-source-id: d7e01c1f3e782694e4600c46d0fdaad590dcba4b