Add WholeProgramState for DexType
Summary:
Add a `WholeProgramState` for the global type analysis that collects the `DexType` for fields and return type for methods at the end of each global iteration. It also feeds the type information into the local type analysis in the next iteration.
At this point it is more of less a copy from the existing `WholeProgramState` from IPCP. Will wire-in the global analyzer in the next diff.

Reviewed By: ssj933

Differential Revision: D19737690

fbshipit-source-id: 76a687ab7c83e4142240afd5de517f506081f125