Abstracting Abstract Interpretation for Interprocedural Analysis
Summary:
Static analysis is hard. Abstract interpretation powered by SPARTA fixpoint iterators can make it easy to reason about properties of graphs. However, interprocedural analysis with abstract interpretation can still be complicated to implement. It's non-trivial to understand what is needed and which component is general among different analyses.

In Redex, we have `IPConstantPropagation` has the setup that exemplifies interprocedural analysis with call graph level fixpoint iteration on SPARTA. In Mariana Trench, we defined the intraprocedural analysis with SPARTA and implemented a global fixpoint iterator in an ad-hoc manner. While we are encouraging more interprocedural analysis on various projects like Redex or LLVM, having a common analysis framework is desired over making copies of existing code.

To address the problem, this diff is proposing `InterproceduralAnalyzer`, a framework that aims to replace the boilerplate code that's required to implement interprocedural analyses. Similar to the rest of SPARTA, the framework utilizes Type arguments to make the code reusable, while maintaining SPARTA's full obliviousness of IR.

The fundamental way of building an analysis with this framework is through providing data structures and transfer functions. The current status requires user to provide a type that represents the function summary, the intraprocedural analysis which gets full access to the summaries, and argument passing domains that can be optionally context sensitive.

The test case in this diff exemplifies pureness analysis on a simple toy language. The intraprocedural analysis can be written in plain C++ without using SPARTA.

I expect this to still go through additional revision. I still have a bunch left TODO. We are landing this for now and plan to iterate on top of it.

Reviewed By: jeremydubreil

Differential Revision: D19592543

fbshipit-source-id: 360ba06cfb7e8960d6ddffa02c63fb45016cefd8