Amend assignability check in IRTypeChecker
Summary:
Android is a somewhat weirdly mixed bag when it comes to checking assignability of interfaces. For dispatch tables, some checking is tighter than the JVM's. For returns, however, it is legal to return any reference type.

This behavior was not imported from Android when the check was added, but is needed for checks to succeed on some inputs.

Reviewed By: int3

Differential Revision: D19916300

fbshipit-source-id: fb49a52d133ab2a025d039bf48484e76ca7ed4a5