sets -> vectors of edges
Summary:
`vector`s are much cheaper to allocate and copy than `set`s. Originally, a `set` was chosen to ensure that we would never create a multi-graph. But I read over the code and realized that we aren't relying on the duplicate removal of the set and won't create a multigraph even with a vector.

The reason is twofold:
* Before any `add_edge` call, we're checking that the registers aren't already aliased
  * `move` checks `are_aliases` before adding any edges
  * `meet_with` checks `are_aliases` before merging groups
* All of our edge removal operations remove all edges, so a multi-graph wouldn't cause a bug anyway
  * `break_alias` uses `clear_vertex` which removes all edges, duplicate or not
  * `join_with` uses `remove_edge(vertex_t u, vertex_t v, ...)` which removes all edges between `u` and `v`

Reviewed By: ssj933

Differential Revision: D19989923

fbshipit-source-id: cea7c7381124b88cdb0f27330d5c4b203accfacc