Introducing PartialPass
Summary: We would like to be able to run certain passes only on select DexStores. By default when the argument is not provided, the behavior of `PartialPass` should be the same as any regular whole program `Passes`. If they are provided, we take over the process of building class scope and subsequently, providing the pass with only the classes it should care about.

Reviewed By: minjang

Differential Revision: D20040297

fbshipit-source-id: 51d4fa2a77d1fd1e1220239f767f325c1a1baea3