Fix PartialPassTest
Summary:
Make the tested pass instance local to the test.

There are two reasons to do this:
* Tests may be performed in parallel.
* If they are performed sequentially, state needs to be reset between runs or take into account previous runs, which, as the test is based on pointers, is dependent on allocator behavior.

Fixes a test divergence under ASAN.

Note: the test can no longer run with a regular PassManager, as passes cannot avoid being registered. We are between a rock (stale memory in the pass registry) and a hard place (leaking instances of previous tests).

Reviewed By: yuxuanchen1997

Differential Revision: D20124480

fbshipit-source-id: 65d7ae634b1f35f1b2ae5e9c347da77908321bfe