bug fix in get_or_create_clinit
Summary:
The way I used to check for clinit existence is wrong. It is possible that clinit gets removed by a previous pass, (like Enum optimization) and when you get_method you can get the old one that isn't attached to the class.

This diff changes the `get_method` existence check to use the DexClass API, so we make sure clinit either exists in the class or an empty one would get created.

Reviewed By: Feng23

Differential Revision: D20316861

fbshipit-source-id: 40f9a55b5a80c8a98188230a2901ae5f33f23a3f