Detect unreachable switch fallthroughs
Summary:
This is an improvement to constant-propagation, enabling it to detect if the fallthrough of a switch is unreachable in case one of the other case-keys matches.

However, it turns out that the constant-propagation pass operates on the *un*editable cfg, where case-keys are not in fact present. Thus, this is a no-nop for this diff. The new functionality only really becomes relevant in D20088729 up the stack where a new unit test depends on it.

Reviewed By: Feng23

Differential Revision: D20091936

fbshipit-source-id: 7f7bde0f171d713155def904edfb555db9e4178b