Ensure field type mapping converges monotonically
Summary:
We had an issue previously where the type mapping of a field in the WholeProgramState can backtrack. The reason is that when propagating the type to local analysis, we sometimes intentionally lower the type if the type domain is top. We did so hoping to increase the number of non-top mappings in the results. However, by doing this, the type mapping of a field can flip flop between a top value and a non-top value between two iterations. We do want to make sure the analysis converges nicely before looking at improving its accuracy. The diff includes the following changes:

- Adjust field type propagation logic so that we don't backtrack a field type mapping between analysis iterations.
- Add debugging logic that prints the difference between an old WholeProgramState and a new one
- Add an integration test for larger input from Java source file

Reviewed By: int3

Differential Revision: D20133989

fbshipit-source-id: cab5e4f2775aa05c3d2ff7a2582393090febbac3