Allow deep_copy to be used on non-fresh CFGs
Summary: Adding a way of clearing everything that is inside a `CFG`. `deep_copy()` does not work unless the `CFG` it is used on is fresh, so if anyone wants to `deep_copy()` inside an existing `CFG` it will not work - hence, provide a `clear()` method and use it inside `deep_copy`.

Reviewed By: justinjhendrick

Differential Revision: D20341392

fbshipit-source-id: 2344d430397558d4e52d78848a16cbd0b193f822