Be more explicit about `Config` lifetime requirement
Summary:
DedupBlocks used to take a `const Config& config` argument and the following code compiles.

```
 DedupBlocks db(Config(), method);
```

In fact the constructor is taking a reference to the `Config()` temporary. This is problematic because nowhere does it require the argument being passed as a const ref to live anywhere.

Changing this to take a `const Config*` instead so it's explicit about the ownership requirement. Alternatively, if we opt to not use raw pointers (or hate the fact that it can be null), we can pass a proper `Config&` instead of `const Config&`.

Reviewed By: budmonde, agampe

Differential Revision: D20402724

fbshipit-source-id: 5bac946eb93a5d3bdf318bc3f8c7421af8924f32