Successfully dereferencing an object implies that it's NEZ
Summary:
The challenge is that we can only apply that change 1) when there's no outgoing throw edge (which is the case for potentially throwing instructions that are not at the end of a block), or, 2) when we are entering a non-throw successor block. A throw successor must not see such a change.

Future work: Determine when throw edges don't actually apply for exception type reasons.

Reviewed By: ssj933

Differential Revision: D20150112

fbshipit-source-id: ddd15dbe4ecd6e2ff27aa932e0df5acbc098dd00