"phi nodes" aka values with multiple definitions
Summary:
Up to now, CSE was only tracking a single definition (=particular instruction) of a value. When two control-flow paths with different definitions were joined, it went to "top'.

This diffs improves that, implementing what would be "phi nodes" in traditional SSA-based CSE implementations. In our case, it simply means that we are tracking sets of definitions, and not just single definitions.

In the transformation, CSE would then allocate one register per set of definitions (*), and store a copy of the defined value for each definition, and use it as before.

(*) We could do with fewer registers, as any subset could use the same register as any of its supersets. I tried that in a diff up the stack,  but it didn’t really do much. Looks like register-allocation/copy-prop efficiently take care of it.

Reviewed By: amnn

Differential Revision: D20431450

fbshipit-source-id: 8723de2a64febc65e1a66387ef2079ddead10bd3