clear cfg after using it
Summary:
The pass builds a (non-editable) cfg that never gets cleared. This is not a problem, as it is non-editable and at the next `build_cfg()` it will be cleared anyways, but it's probably better to do it in the service.

I also tried to put up a diff with an editable CFG, but it requires more changes.

Reviewed By: NTillmann

Differential Revision: D20675118

fbshipit-source-id: 9eff06c2125b6a0aeccf70bee6f283ed3b41f264