Split huge switches
Summary:
Add a pass to split large methods with a big switch. Add a test.

The switch expression must be simple and side-effect-free. The split is simplistic:
* The method is replicated for each split, and the corresponding cases removed.
* An `if` chain on entry dispatches to a split, if necessary.

The switch splits are not size-driven. A range as determined by the number of splits is grabbed. This simplistic approach seems good enough in practice for now, but can be replaced with a more expensive grab-and-retry later.

Splits are only beneficial for runtime performance if the splits are compiled or the common cases are in the first split, as each call exhibits overhead (compared to the table jump for `packed-switch` or binary search for `sparse-switch`). Add a method whitelist filter so that some methods can be skipped.

If method stats are available, restrict splitting to "hot" methods (as configurable by threshold), so as to not increase code size without potential performance gain.

Reviewed By: NTillmann

Differential Revision: D20084557

fbshipit-source-id: 7bb927c198a8cffb0389553f964003f056210093