Properly erase the enum construction code in <clinit>
Summary:
The pass wants to erase the enum construction code and replace all the other enum references with integer type.
It relies on specific pattern to erase the construction code without considering that some of the enum object registers may be alive after the normal enum construction finishes. Fix it.

See the comments for the example code transformation.

TODO: add unittest

Reviewed By: NTillmann

Differential Revision: D20909452

fbshipit-source-id: d1705d51e0d4ddaa3b183b747f5728cd7484eaf5