Fix LocalPointersAnalysis leak
Summary:
Avoid leaking fixpoint iterators when the map already contains an entry.

This is a quick fix - make iterator map update (freeing an old value if it exists) and summary map update a critical section. Both must be together as otherwise a racing thread may free the iterator of the first before it can generate the summary.

The critical section approach was taken so that the summary map always corresponds to the iterator in the map. That may be helpful in the case of later iterators (different call path) having more information (callgraph loops). Alternatives are to always only keep the first entry, or to use a `shared_ptr` in the iterator map and let the summary map "diverge."

Reviewed By: justinjhendrick

Differential Revision: D20845998

fbshipit-source-id: 17041bc0b81e2d155a261a41c3fb15b98ae3fa86