use trees instead of cliques in AliasedRegisters
Summary:
AliasedRegisters is a data structure that CopyPropagation uses to keep track of which Values (registers, constants, final fields, etc.) are the same (aliased).

Under the hood, AliasedRegisters uses a graph to keep track of these alias relationships. Before this diff, it would create "cliques" (fully connected subgraphs) for each group of aliased values. But the problem was that it used O(N^2) number of edges. On large methods this became a very large use of memory.

This diff replaces the cliques with trees, similar to a [link/cut tree](https://en.wikipedia.org/wiki/Link/cut_tree), but these trees only have two levels (every node is either a root or a leaf). These trees fill the same use case as the cliques (if two nodes are in the same tree, then they are aliased) but with much less memory use. The number of edges is O(N) instead of O(N^2).

A single clique could be represented as multiple different trees (by choosing different roots), but this is undesirable. To enforce canonical trees, this diff introduces an `operator<` to the Values (nodes) so that the minimum node is always the root node. This incurs the cost of changing the root node more frequently, but it's worth it over all, especially when computing the intersection of two graphs.

Reviewed By: int3

Differential Revision: D20684226

fbshipit-source-id: 279313a7b16adb8ceb90cf36926a320d5d5df9f1