introduce weakening
Summary:
Track "type demands", and weaken the type we are casting to. This may cause a cast to get removed/replaced, if weakened type is implied, or we may rewrite the cast to the weaker type.

Let's say we have the following.
```
class C extends B {}
class B extends A { public int F; }
class A {}
```
We'll do  **in-place cast weakening** and **cast removal after weakening**. For example
```
int foo(A a) {
  C c = (C)a; // cast stronger than demanded by upcoming field access!
  return c.f;
}
int bar(B b) {
  C c = (C)b; // cast stronger than demanded by upcoming field access
  return c.f;
}
```
would get rewritten to
```
int foo(A a) {
  B b = (B)b;
  return b.f;
}
int bar(B b) {
  return b.f;
}
```
The new optimization is guarded by a new `RemoveRedundantCheckCasts` config option `weaken` that's on by default.

This optimization has a small potential to cause problems if the code relies on catching `ClassCastException`. Note that the `LocalDCE` pass already does something similar: It removes check-cast instructions if their result is not used.

Reviewed By: amnn

Differential Revision: D20829516

fbshipit-source-id: 0b8ec703a7d36bcbec3b9ca10efd5601349448a6