Synchronize changes in ProguardMatcher
Summary:
Multiple keep rules can affect the same class. As the `ReferenceState` flags are not atomic, and especially since they are now bit flags, avoid races.

Avoid the memory and complexity overhead of a lock per class. Instead simplify the implementation by using an array of mutexes for sharding of the locks.

Overhead is negligible.

Reviewed By: minjang

Differential Revision: D20976007

fbshipit-source-id: a94150d842f7e7a86e8a60983eff918800cc1a25