Improve DexTypeValue join operation
Summary:
The join operation of `DexTypeValue` used to give up if the direct parents of the two operands are not a subclass of the other. This is incomplete. For a very steep hierarchy, we weren't able to find the common parent and join to it.

This change adds the full hierarchy scan on both operand all the way to Ljava/lang/Object; to find the common parent. The improvement paves the way for the incoming changes on `SmallSetDexTypeDomain` built on top of `DexTypeValue`.

Reviewed By: yuxuanchen1997

Differential Revision: D20852598

fbshipit-source-id: 74204b4dcbe4523beba6a993d385641e38ea257b