Create a `DexDebugItem` when inlining if necessary
Summary:
If code with debug info is inlined into code without debug info, create a `DexDebugItem` so that inlined positions can be encoded.

Note that this happens only for `inliner` functions, not `CFGInliner` directly, as context is missing.

Use the class' source file if available. As we do not currently support `Unknown Source` correctly, encode the file as `UnknownSource` when there is no hint. Use line number `0`.

Update tests.

Reviewed By: justinjhendrick

Differential Revision: D21079040

fbshipit-source-id: f1886584c878aaab4ea7a0c9b76072faf7a7ac53