Tolerate constructor references that have internal types in their signatures without def
Summary:
For constructor methods that have no method definitions, we still update their signatures properly.

About `TypeRefUpdater`,
it is for replacing internal type refernces with other type references in signatures of methods and fields.
If we want to update `Foo` with `Bar` for methods, we may get
`A.method:(LFoo;)V` => `A.method$REDEX$xxx:(LBar;)V`
`A.<init>:(LFoo;)V` = `A.<init>:(LBar;)V` or `A.<init>:(LBar;IIIIIII)V`

For `<init>` methods, we should never rename them.

Reviewed By: emma0303

Differential Revision: D21161807

fbshipit-source-id: 7bc9e98d9a898482890d836bcee472d16f9a864e