Make Redex pretty printing a custom command
Summary: If the data structure which we are printing is inconsistent, it could lead to gdb crashing while debugging Redex. If we use the standard pretty printing infrastructure, gdb can call the pretty printer even when we break, not only when we print. Thus proposing a custom print command "pp".

Reviewed By: amnn

Differential Revision: D21085932

fbshipit-source-id: ab4002703208e196f18c92d87cc41e5237a52ecf