Simplify SmallSetDexTypeDomain
Summary:
The initial implementation of `SmallSetDexTypeDomain` is too slow once plugged into `TypeInference`. The `join` operation is too expensive, and it is the most frequently perform operation. We cannot afford a complex `join` operation.

The new design put the small `PatriciaTreeSet` as an additional product of the `DexTypeDomain` along side the single type domain we currently operate on. This approach enables us to join to Top faster, so as to avoid unwanted `join` operation that's expensive to perform. Both the single type domain as well as the set type domain are available at the end of the analysis. `TypeInference` doesn't show conceivable performance degradation.

Reviewed By: Feng23

Differential Revision: D20979924

fbshipit-source-id: 331c1973a3c887481440e709987e3a8cbfa0303d