Track whether allocators have been reached in the store
Summary:
When adding the test cases included in the diff, it emerged that the analysis
was not able to effectively track allocations that were not always reached:

```
B0:   load-param v0
      if-nez v0 :B2
B1: * new-instance LFoo;
      move-result-pseudo-object v0
      invoke-direct v0 LFoo;.<init>:()V
B2:   return-object v0
```

Creating a "control-flow triangle" with the initial instructions of each
basic-block annotated with labels (suspending disbelief for the moment that
`new-instance` might throw an exception).

We are interested in tracking the escapes for the `*`-ed instruction.

At the end of `B0`, the abstract heap (store) is empty, meaning all abstract
values map to Top, and at the end of `B1`, for `*`, we know that it may
have escaped once (by the `invoke-direct`).

Before analysing `B2` the abstract states from the analyses of `B0` and `B1`
(its incident edges) will need to be joined.  Unfortunately, this join discards
the more precise information because at `*` it is joined with Top (resulting in
Top).

This diff gets around this problem by "lifting" the values in the abstract
heap, so they now have two levels of least element.  Bottom, meaning the
analysis is in some inconsistent state (this one participates in smashing), and
the Lifted Bottom, which in this case represents an allocation that hasn't
happened yet.  Before starting, the store is seeded with a lifted bottom for
each allocation site we are interested in tracking.

In short while previously, the analysis could assume nothing about unallocated
values, now it can assume that unallocated values contain nothing!  This
provides the fidelity we need to track allocations that do not dominate the
exit block.

Reviewed By: kathryngray

Differential Revision: D21227707

fbshipit-source-id: 95672f70ad707d8667ca11cdadabf0f367133ab5