Fix edge case of blaming escape analysis
Summary:
Previously if an `invoke` was treated as an allocator, its parameters would not
be treated as escaping, and if it were treated as safe, this would apply
equally to its parameters and its return value.

This was too coarse to be generally useful -- now, instead, safe method
invocations are considered not to escape their parameters and invocations
treated as allocators are assumed not to escape the returned instance, but
otherwise the invoke is assumed to escape its inputs/output.

Reviewed By: int3

Differential Revision: D21159137

fbshipit-source-id: f65d3e39f2f3ecc1dadb3db535af43900c057b18