feat(website): add DocSearch as recommended by docusaurus (#472)
Summary:
👋 team,

I am working on DocSearch. [We have an integration for Docusaurus](https://docusaurus.io/docs/en/search#enabling-the-search-bar). We thought it is a shame you can not also used this search that you can [find on reactjs for example](https://reactjs.org/).

This PR will add DocSearch to the documentation website. It will allow an user to have a learn-as-you-type experience by displaying results thanks to a dropdown in a live way.

Let me know if you need anything.
Pull Request resolved: https://github.com/facebook/redex/pull/472

Reviewed By: int3

Differential Revision: D21261973

Pulled By: justinjhendrick

fbshipit-source-id: 3e4f0ea35aa5019d9bd94542bdfeefca43464dbc