Introducing preserved analysis
Summary:
LLVM has this notion of preserved analysis. Considering that some analyses we are implementing can take a while to run, this can be useful.

However, the functionality is rather simple, the pass manager does not schedule to run an analysis pass by itself. It still relies on the config. However, checks are being added to make sure the required analyses are run and haven't been invalidated.

Reviewed By: thezhangwei

Differential Revision: D20236636

fbshipit-source-id: f379bf3cf9ae70a4eae69a39cf450aa6fbbeff19