Avoid copy of container for `in` combinator.
Summary:
The `m::in` combinator takes a collection, and creates a matcher which succeeds
when the passed in value is in the container (according to `find`).  Previously
it copied the collection in question in order to own it, but to my mind this
isn't so ideal -- I had a use case (to come in a future diff) that would have
resulted in this combinator frequently copying quite a heavy object.

I expect the resulting matcher to be created just before use, used and then
discarded, and so it is more appropriate for it to borrow the collection than
own it.  The existing use cases (I could find only one) seem to follow this
assumption.

Reviewed By: newobj

Differential Revision: D21331482

fbshipit-source-id: 7942e1ea7c254c62ee4c5465b570aca116c3e97d